Created some basic locking functionality
Added a UI window to control which lock variable gets assigned to which
lock. Had to create some basic logic that only allows one variable to be
selected at a time. Then I created a unity event for each lock which
will call all locks that are currently assigned to it.

My next step is to set up linking locks and target nodes.